remove puppeteer utils
