+++ javascript/test/URDFLoader.test.js
@@ -228,6 +228,66 @@ describe('Load', () => {
 
 });
 
+describe('Material Tags', () => {
+
+    it('should parse material colors and name.', () => {
+
+        const loader = new URDFLoader();
+        const res = loader.parse(`
+            <robot name="TEST">
+                <material name="Cyan">
+                    <color rgba="0 1.0 1.0 1.0"/>
+                </material>
+                <link name="LINK">
+                    <visual>
+                        <geometry>
+                            <box size="1 1 1"/>
+                        </geometry>
+                        <material name="Cyan"/>
+                    </visual>
+                </link>
+            </robot>
+        `);
+
+        const material = res.children[0].children[0].material;
+        expect(material.name).toEqual('Cyan');
+        expect(material.color).toEqual(new Color(0, 1, 1));
+        expect(material.transparent).toEqual(false);
+        expect(material.depthWrite).toEqual(true);
+        expect(material.opacity).toEqual(1.0);
+
+    });
+
+    it('should parse transparent materials correctly.', () => {
+
+        const loader = new URDFLoader();
+        const res = loader.parse(`
+            <robot name="TEST">
+                <material name="Cyan">
+                    <color rgba="0 1.0 1.0 0.5"/>
+                </material>
+                <link name="LINK">
+                    <visual>
+                        <geometry>
+                            <box size="1 1 1"/>
+                        </geometry>
+                        <material name="Cyan"/>
+                    </visual>
+                </link>
+            </robot>
+        `);
+
+        const material = res.children[0].children[0].material;
+        expect(material.name).toEqual('Cyan');
+        expect(material.color).toEqual(new Color(0, 1, 1));
+        expect(material.transparent).toEqual(true);
+        expect(material.depthWrite).toEqual(false);
+        expect(material.opacity).toEqual(0.5);
+
+    });
+
+});
+
 describe('TriATHLETE Climbing URDF', () => {
 
     let robot;
